requirements.txt
aiogram==3.6.0
python-dotenv
